No commit message

+++ scripts/r/linux/setup_awesomewm.sh
@@ -1,13 +1,15 @@
 set -e
 
 # Install required packages
-packages="awesome xbacklight alsa-utils"
-for package in $packages; do
-    dpkg -s "$package" >/dev/null 2>&1 || {
-        sudo apt-get update
-        sudo apt-get install -y "$package"
-    }
-done
+if [[ -f "/etc/debian_version" ]]; then
+    packages="awesome xbacklight alsa-utils"
+    for package in $packages; do
+        dpkg -s "$package" >/dev/null 2>&1 || {
+            sudo apt-get update
+            sudo apt-get install -y "$package"
+        }
+    done
+fi
 
 # Copy awesome config file
 ln -sf "$(dirname "$0")/../../../settings/awesome" $HOME/.config/
