+++ scripts/r/linux/setup_linux_surface.sh
@@ -1,15 +1,18 @@
 set -e
 
+# Install surface linux kernel
 if [[ "$(uname -r)" != *"-surface"* ]]; then
-    # https://github.com/linux-surface/linux-surface/wiki/Installation-and-Setup#manually-installing-the-repository
-    wget -qO - https://raw.githubusercontent.com/linux-surface/linux-surface/master/pkg/keys/surface.asc |
-        gpg --dearmor | sudo dd of=/etc/apt/trusted.gpg.d/linux-surface.gpg
-    echo "deb [arch=amd64] https://pkg.surfacelinux.com/debian release main" |
-        sudo tee /etc/apt/sources.list.d/linux-surface.list
-    sudo apt update
-    sudo apt install linux-image-surface linux-headers-surface libwacom-surface iptsd
-    # sudo apt install linux-surface-secureboot-mok -y
-    sudo update-grub
+    if [[ -f "/etc/debian_version" ]]; then
+        # https://github.com/linux-surface/linux-surface/wiki/Installation-and-Setup#manually-installing-the-repository
+        wget -qO - https://raw.githubusercontent.com/linux-surface/linux-surface/master/pkg/keys/surface.asc |
+            gpg --dearmor | sudo dd of=/etc/apt/trusted.gpg.d/linux-surface.gpg
+        echo "deb [arch=amd64] https://pkg.surfacelinux.com/debian release main" |
+            sudo tee /etc/apt/sources.list.d/linux-surface.list
+        sudo apt update
+        sudo apt install linux-image-surface linux-headers-surface libwacom-surface iptsd
+        # sudo apt install linux-surface-secureboot-mok -y
+        sudo update-grub
+    fi
 fi
 
 # Config thermald
@@ -20,7 +23,11 @@ sudo systemctl restart thermald.service
 sudo systemctl status thermald.service
 
 # Brightness control
-sudo apt install light -y
+if command -v apt &>/dev/null; then
+    sudo apt install light -y
+else
+    sudo pacman -S light --noconfirm
+fi
 sudo usermod -aG sudo $(whoami)
 sudo tee /etc/udev/rules.d/90-brightnessctl.rules <<-EOF
 ACTION=="add", SUBSYSTEM=="backlight", RUN+="/bin/chgrp video /sys/class/backlight/%k/brightness"
